Everything is converted
Changes:
 1. added function to EmbedManager.java that allows one to turn an EmbedBuilder object into an arraylist of one element. Useful for submenus
 2. Modified enterSubmenu to require a pageNum. Allows for the method to do things based upon the previously displayed page.
 3. moved SearchPaginator.java & SearchSubmenu2.java over to a new searchPaginators package
 4. Deleted old paginator classes SearchEmbedMenu.java & SearchSubmenu.java
 5. Refactored SearchSubmenu2.java to SearchSubmenu.java
