+++ src/main/java/paginators/searchPaginators/SearchPaginator.java
@@ -1,11 +1,14 @@
-package paginators.pWithSubmenu;
+package paginators.searchPaginators;
 
+import apis.ombi.OmbiCallers;
+import apis.ombi.templateClasses.movies.moreInfo.MovieInfo;
+import apis.ombi.templateClasses.tv.moreInfo.TvInfo;
 import com.jagrosh.jdautilities.command.CommandEvent;
 import com.jagrosh.jdautilities.commons.waiter.EventWaiter;
 import com.jagrosh.jdautilities.menu.Menu;
+import discordBot.EmbedManager;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Message;
-import net.dv8tion.jda.api.entities.MessageChannel;
 import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.message.react.MessageReactionAddEvent;
@@ -14,6 +17,7 @@ import net.dv8tion.jda.internal.utils.Checks;
 import paginators.Paginator;
 
 import java.util.ArrayList;
+import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
@@ -24,9 +28,16 @@ import java.util.function.Consumer;
 //                      "\u2705",        //✅
 //                      "\u25B6"}        //▶️️
 
+
 public class SearchPaginator extends Paginator {
-    public SearchPaginator(EventWaiter waiter, Set<User> users, Set<Role> roles, long timeout, TimeUnit timeUnit, ArrayList<EmbedBuilder> pages, boolean wrapPages, Consumer<Message> finalAction) {
+
+    //add a few variables pertaining to media type
+    private final int MEDIA_TYPE;
+
+
+    public SearchPaginator(EventWaiter waiter, Set<User> users, Set<Role> roles, long timeout, TimeUnit timeUnit, ArrayList<EmbedBuilder> pages, boolean wrapPages, Consumer<Message> finalAction, int mediaType) {
         super(waiter, users, roles, timeout, timeUnit, new String[]{":arrow_backward:", ":stop_sign:", ":white_check_mark:", ":arrow_forward:"}, pages, wrapPages, finalAction);
+        MEDIA_TYPE = mediaType;
     }
 
     @Override
@@ -48,7 +59,7 @@ public class SearchPaginator extends Paginator {
 
         } else if (REACTIONS[2].getUnicode().equals(event.getReaction().getReactionEmote().getName())) { //the third item in the array is the check mark
             //Enter the submenu
-            enterSubmenu(event.getChannel());
+            enterSubmenu(message, pageNum);
 
         } else if (REACTIONS[3].getUnicode().equals(event.getReaction().getReactionEmote().getName())) { //the last item is the right arrow
             //Increment pageNum
@@ -69,8 +80,48 @@ public class SearchPaginator extends Paginator {
     }
 
     @Override
-    protected void enterSubmenu(MessageChannel channel) {
-        System.out.println("not a thing yet!");
+    protected void enterSubmenu(Message oldMessage, int pageNum) {
+        //Create SubmenuPaginator
+        OmbiCallers caller = new OmbiCallers();
+        EmbedManager embedManager = new EmbedManager();
+        SearchSubmenu.Builder submenuBuilder = new SearchSubmenu.Builder()
+                .setFinalAction(message -> {
+                    try {
+                        message.clearReactions().queue();
+                    } catch (PermissionException ex) {
+                        message.delete().queue();
+                    }
+                })
+                .setEventWaiter(waiter);
+        //Set Media type
+        if (MEDIA_TYPE == 1) {
+
+            //extract the media ID from the tvEmbed
+            int mediaID = Integer.parseInt(Objects.requireNonNull(EMBED_ARRAYLIST.get(pageNum).getFields().get(0).getValue(), "Media ID is null! Impossible to continue!"));
+
+            //get the media embed and the status
+            TvInfo info = caller.ombiTvInfo(String.valueOf(mediaID));
+
+
+            submenuBuilder.setTvInfo(info);
+            submenuBuilder.setEmbedArray(embedManager.toArrayList(embedManager.createTvMoreInfoEmbed(info)));
+            submenuBuilder.setTvType();
+        } else if (MEDIA_TYPE == 2) {
+            //extract the media ID from the tvEmbed
+            int mediaID = Integer.parseInt(Objects.requireNonNull(EMBED_ARRAYLIST.get(pageNum).getFields().get(0).getValue(), "Media ID is null! Impossible to continue!"));
+
+            //we need to generate the MovieInfo object in order to pass it to the paginator and get the required embedArray
+            MovieInfo info = caller.ombiMovieInfo(String.valueOf(mediaID));
+
+            submenuBuilder.setMovieType();
+            submenuBuilder.setEmbedArray(embedManager.toArrayList(embedManager.createMovieMoreInfoEmbed(info)));
+            submenuBuilder.setMovieInfo(info);
+        }
+        SearchSubmenu p = submenuBuilder
+                .build();
+
+        p.paginate(oldMessage, 1);
+
     }
 
     public static class Builder extends Menu.Builder<SearchPaginator.Builder, SearchPaginator> {
@@ -89,7 +140,7 @@ public class SearchPaginator extends Paginator {
             Checks.check(waiter != null, "Must set an EventWaiter");
             Checks.check(!embeds.isEmpty(), "Must include at least one item to paginate");
 
-            return new SearchPaginator(waiter, users, roles, timeout, unit, embeds, wrapPageEnds, finalAction);
+            return new SearchPaginator(waiter, users, roles, timeout, unit, embeds, wrapPageEnds, finalAction, type);
 
         }
 
@@ -131,4 +182,5 @@ public class SearchPaginator extends Paginator {
             this.wrapPageEnds = wrapPageEnds;
         }
     }
+
 }
